Рефакторинг кода

+++ WordsParser.Infrastructure/Configurations/FileSettings.cs
@@ -4,5 +4,5 @@ namespace WordsParser.Infrastructure.Configurations;
 
 public class FileSettings : IFileSettings
 {
-    public int MaxFileSizeMbytes { get; set; }
+    public decimal MaxFileSizeMbytes { get; set; }
 }
